+++ CHANGELOG.rst
@@ -8,6 +8,15 @@ Contributors to this version: Juliette Lavoie (:user:`juliettelavoie`), Pascal B
 
 New features and enhancements
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+* The `mask` argument in ``stack_drop_nans`` can now be a list of dimensions. In that case, a `dropna(how='all')` operation will be used to create the mask on-the-fly. (:pull:`450`).
+* Few changes to ``clean_up``:
+    * The `convert_calendar` function now uses `xarray` instead of `xclim`. (:pull:`450`).
+    * The `attrs_to_remove` and `remove_all_attrs_except` arguments now use real regex. (:pull:`450`).
+    * Multiple entries can now be given for `change_attr_prefix`. (:pull:`450`).
+* ``minimum_calendar`` now accepts a list as input. (:pull:`450`).
+* More calendars are now recognized in ``translate_time_chunk``. (:pull:`450`).
+* `new_dim` in ``unstack_dates`` is now None by default and changes depending on the frequency. It becomes `month` if the data is exactly monthly, and keep the old default of `season` otherwise. (:pull:`450`).
+* Updated the list of libraries in `show_versions` to reflect our current environment. (:pull:`450`).
 * New ``xscen.catutils.patterns_from_schema`` to generate all possible patterns from a given schema (or one of xscen's default), to use with :py:func:`parse_directory`. (:pull:`431`).
 * New ``DataCatalog.copy_files`` to copy all files of catalog to a new destination, unzipping if needed and returning a new catalog. (:pull:`431`).
 * Convenience functions ``xs.io.zip_directory`` and ``xs.io.unzip_directory`` (for zarrs). (:pull:`431`).
@@ -21,6 +30,10 @@ Bug fixes
 * Avoid modification of mutable arguments in ``search_data_catalogs`` (:pull:`413`).
 * ``ensure_correct_time`` now correctly handles cases where timesteps are missing. (:pull:`440`).
 * If using the argument `tile_buffer` with a `shape` method in ``spatial.subset``, the shapefile will now be reprojected to a WGS84 grid before the buffer is applied. (:pull:`440`).
+* ``maybe_unstack`` now works if the dimension name is not the default. (:pull:`450`).
+* ``unstack_fill_nan`` now works if given a dictionary that contains both dimensions and coordinates. (:pull:`450`).
+* ``clean_up`` no longer modifies the original dataset. (:pull:`450`).
+* ``unstack_dates`` now works correctly for yearly datasets when `winter_starts_year=True`, as well as multi-year datasets. (:pull:`450`).
 * Fix ``xs.catalog.concat_data_catalogs`` for catalogs that have not been search yet. (:pull:`431`).
 * Fix indicator computation using ``freq=2Q*`` by assuming this means a semiannual frequency anchored at the given month (pandas assumes 2 quarter steps, any of them anchored at the given month). (:pull:`431`).
 
@@ -32,6 +45,12 @@ Internal changes
 * Add ``.zip`` and ``.zarr.zip`` as possible file extensions for Zarr datasets. (:pull:`426`).
 * Explicitly assign coords of multiindex in `xs.unstack_fill_nan`. (:pull:`427`).
 * French translations are compiled offline. A new check ensures no PR are merged with missing messages. (:issue:`342`, :pull:`443`).
+* Continued work to add tests. (:pull:`450`).
+
+Breaking changes
+^^^^^^^^^^^^^^^^
+* `convert_calendar` in ``clean_up`` now uses `xarray` instead of `xclim`. Keywords aren't compatible between the two, but given that `xclim` will abandon its function, no backwards compatibility was sought. (:pull:`450`).
+* `attrs_to_remove` and `remove_all_attrs_except` in ``clean_up`` now use real regex. It should not be too breaking since a `fullmatch()` is used, but `*` is now `.*`. (:pull:`450`).
 
 v0.9.1 (2024-06-04)
 -------------------